Refactored and Documented init.gd and boot.tscn
I went through and refactored some of the init.gd code to be cleaner, as
well as use proper variable references. I did the same with the scene
itself, renaming things to be clearer. The only major code changes was
the extra loop, and some dummy comments I decided to remove.
main.gd also got a bit of documentation as well.